rename Bulma component folder
