+++ src/components/Custom/LabeledProgressBar.scss
