+++ src/components/TaskAnalysisTable/TaskAnalysisTableHeader.js
@@ -9,7 +9,7 @@ import _omit from 'lodash/omit'
 import AsManager from '../../interactions/User/AsManager'
 import Dropdown from '../Dropdown/Dropdown'
 import SvgSymbol from '../SvgSymbol/SvgSymbol'
-import TriStateCheckbox from '../Bulma/TriStateCheckbox'
+import TriStateCheckbox from '../Custom/TriStateCheckbox'
 import ConfirmAction from '../ConfirmAction/ConfirmAction'
 import TimezonePicker from '../TimezonePicker/TimezonePicker'
 import PagePicker from '../PagePicker/PagePicker'
