+++ src/components/TaskPane/ActiveTaskDetails/KeyboardShortcutReference/KeyboardShortcutReference.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import classNames from 'classnames'
 import { FormattedMessage, injectIntl } from 'react-intl'
 import _omit from 'lodash/omit'
-import Popout from '../../../Bulma/Popout'
+import Popout from '../../../Custom/Popout'
 import SvgSymbol from '../../../SvgSymbol/SvgSymbol'
 import KeyboardShortcutList from './KeyboardShortcutList'
 import messages from './Messages'
